+++ src/hapi/pipelines/app/pipelines.py
@@ -19,6 +19,7 @@ from hapi.pipelines.database.operational_presence import OperationalPresence
 from hapi.pipelines.database.org import Org
 from hapi.pipelines.database.org_type import OrgType
 from hapi.pipelines.database.population import Population
+from hapi.pipelines.database.refugees import Refugees
 from hapi.pipelines.database.sector import Sector
 
 
@@ -153,6 +154,7 @@ class Pipelines:
             "humanitarian_needs", "admintwo", adminlevel=self.admintwo
         )
         _create_configurable_scrapers("national_risk", "national")
+        _create_configurable_scrapers("refugees", "national")
 
     def run(self):
         self.runner.run()
@@ -239,3 +241,16 @@ class Pipelines:
                 results=results,
             )
             national_risk.populate()
+
+        if not self.themes_to_run or "refugees" in self.themes_to_run:
+            results = self.runner.get_hapi_results(
+                self.configurable_scrapers["refugees"]
+            )
+
+            refugees = Refugees(
+                session=self.session,
+                metadata=self.metadata,
+                locations=self.locations,
+                results=results,
+            )
+            refugees.populate()