+++ src/hapi/pipelines/database/humanitarian_needs.py
@@ -7,6 +7,10 @@ from hapi_schema.db_humanitarian_needs import DBHumanitarianNeeds
 from hxl.model import Column, TagPattern
 from sqlalchemy.orm import Session
 
+from ..utilities.parse_tags import (
+    get_gender_and_age_range,
+    get_min_and_max_age,
+)
 from . import admins
 from .base_uploader import BaseUploader
 from .metadata import Metadata
@@ -56,10 +60,9 @@ class HumanitarianNeeds(BaseUploader):
                     if not sector_code:
                         sector_code = "*"
                     sector_code = sector_code.upper()
-                    gender = _get_gender(column)
                     # "#*+age0_4" "#*+age80plus"
-                    age_range = _get_age_range(hxl_tag)
-                    min_age, max_age = _get_min_and_max_age(age_range)
+                    gender, age_range = get_gender_and_age_range(hxl_tag)
+                    min_age, max_age = get_min_and_max_age(age_range)
                     # "#*+disabled"
                     disabled_marker = _get_disabled_marker(column)
                     # TODO: Will there be columns for able bodied?
@@ -110,16 +113,6 @@ def _get_population_status(col: Column) -> str:
     return population_status
 
 
-def _get_gender(col: Column) -> str:
-    gender_patterns = {
-        TagPattern.parse(f"#*+{g}"): g for g in ["f", "m", "x", "u", "o", "e"]
-    }
-    gender = match_column(col, gender_patterns)
-    if not gender:
-        gender = "*"
-    return gender
-
-
 def _get_population_group(col: Column) -> str:
     population_group_patterns = {
         TagPattern.parse("#*+refugees"): "REF",
@@ -132,37 +125,9 @@ def _get_population_group(col: Column) -> str:
     return population_group
 
 
-def _get_age_range(hxl_tag: str) -> str:
-    age_component = hxl_tag.split("+")[-1]
-    age_range = "*"
-    if not age_component.startswith("age"):
-        return age_range
-    age_component = age_component[3:]
-    if age_component.endswith("plus"):
-        age_range = age_component[:-4] + "+"
-    else:
-        age_range = age_component.replace("_", "-")
-    return age_range
-
-
 def _get_disabled_marker(col: Column) -> str:
     disabled_marker = TagPattern.parse("#*+disabled").match(col)
     if disabled_marker:
         return "y"
     if not disabled_marker:
         return "*"
-
-
-# TODO: this is duplicate code, either move to shared location or overwrite with new HNO pipeline
-def _get_min_and_max_age(age_range: str) -> (int | None, int | None):
-    if age_range == "*":
-        return None, None
-    ages = age_range.split("-")
-    if len(ages) == 2:
-        # Format: 0-5
-        min_age, max_age = int(ages[0]), int(ages[1])
-    else:
-        # Format: 80+
-        min_age = int(age_range.replace("+", ""))
-        max_age = None
-    return min_age, max_age